--- packages/react-dnd/src/hooks/useRefObject.ts
@@ -1,9 +1,0 @@
-import { RefObject, useRef, useEffect } from 'react'
-
-export function useRefObject<T>(input: T): RefObject<T> {
-	const ref = useRef(input)
-	useEffect(() => {
-		ref.current = input
-	}, [input])
-	return ref
-}
